Another hack, the last one added bga images support... this add video, we have bga video support. Thanks to the vlcj project and the VLC project.
We can't distribute the vlc files because licences issues and lazyness (D:)
In the Configuration tab you can select it, don't forget to save after.
Also:
If you are running a 64-bit JVM, then you need a 64-bit build of vlc.
If you are running a 32-bit JVM, then you need a 32-bit build of vlc.